requirements.txt
ultralytics>=8.0.0
gradio==4.44.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
Flask>=3.0.0
openai>=1.40.0
python-dotenv>=1.0.0
twilio>=8.0.0